chore(space-wizard): refactor/rename/move of some files
